Move the RemoteEvent value object into the Webhook namespace
src/RemoteEvent/RemoteEvent.php -> src/Webhook/RemoteEvent.php
(Setono\SyliusShipmondoPlugin\RemoteEvent\RemoteEvent ->
 Setono\SyliusShipmondoPlugin\Webhook\RemoteEvent); it is only used by the
webhook parser/consumer/handlers, so it belongs next to them.
