+++ src/Webhook/Handler/CompositeRemoteEventHandler.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Setono\SyliusShipmondoPlugin\Webhook\Handler;
 
 use Setono\CompositeCompilerPass\CompositeService;
-use Setono\SyliusShipmondoPlugin\RemoteEvent\RemoteEvent;
+use Setono\SyliusShipmondoPlugin\Webhook\RemoteEvent;
 
 /**
  * @extends CompositeService<RemoteEventHandlerInterface>
