+++ src/Webhook/Handler/PaymentStateHandler.php
@@ -7,8 +7,8 @@ namespace Setono\SyliusShipmondoPlugin\Webhook\Handler;
 use Doctrine\Persistence\ManagerRegistry;
 use Setono\Doctrine\ORMTrait;
 use Setono\SyliusShipmondoPlugin\Model\OrderInterface;
-use Setono\SyliusShipmondoPlugin\RemoteEvent\RemoteEvent;
 use Setono\SyliusShipmondoPlugin\Webhook\OrderResolverInterface;
+use Setono\SyliusShipmondoPlugin\Webhook\RemoteEvent;
 use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Component\Core\OrderPaymentTransitions;
 use Sylius\Component\Payment\PaymentTransitions;
