+++ tests/Unit/Webhook/Handler/FulfillOrderHandlerTest.php
@@ -12,9 +12,9 @@ use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Setono\SyliusShipmondoPlugin\Model\OrderInterface;
-use Setono\SyliusShipmondoPlugin\RemoteEvent\RemoteEvent;
 use Setono\SyliusShipmondoPlugin\Webhook\Handler\FulfillOrderHandler;
 use Setono\SyliusShipmondoPlugin\Webhook\OrderResolverInterface;
+use Setono\SyliusShipmondoPlugin\Webhook\RemoteEvent;
 use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Order\StateResolver\StateResolverInterface;
